% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{bind_cols_fill}
\alias{bind_cols_fill}
\title{bind_cols_fill}
\usage{
bind_cols_fill(lst)
}
\arguments{
\item{lst}{a list of vectors}
}
\value{
tibble
}
\description{
takes a list of vectors of potentilly differeing lengths and binds them 
together as columns in a tibble filling any empty values with NAs
}
