% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{dge_filters}
\alias{dge_filters}
\title{dge_filters}
\usage{
dge_filters(data, dir_filter, padjt, lowerlfc, upperlfc)
}
\arguments{
\item{data}{results_annotated_min_cov_grp}

\item{dir_filter}{dplyr function function for a particualt direction}

\item{padjt}{minimum adjusted p-value}

\item{lowerlfc}{lower log2 Fold Change cut-off}

\item{upperlfc}{upper log2 Fold Change cut-off}
}
\value{
filtered tibble
}
\description{
applies filters for adjusted p-value, direction and log2 Fold Change to 
differential gene expression data
}
