% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{format_results}
\alias{format_results}
\title{format_results}
\usage{
format_results(res_extra, organism_name)
}
\arguments{
\item{res_extra}{the extended results table from which to generate the new columns}

\item{organism_name}{latin name of the organism, used to make links to ensembl}
}
\value{
tibble
}
\description{
format_results
}
