% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{gen_df_DT_tibble}
\alias{gen_df_DT_tibble}
\title{gen_df_DT_tibble}
\usage{
gen_df_DT_tibble(
  significant_genes_by_comparison,
  intersection_selected_sets,
  species
)
}
\arguments{
\item{significant_genes_by_comparison}{significant_genes_by_comparison}

\item{intersection_selected_sets}{intersection_selected_sets}

\item{species}{The latin binomial name of the species (used to make ensembl links)}
}
\value{
a nested tibble with columns for comparisons, data and DT objects as named lists
}
\description{
gen_df_DT_tibble
}
