% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{gen_DT_lst}
\alias{gen_DT_lst}
\title{gen_DT_lst}
\usage{
gen_DT_lst(
  significant_genes_by_comparison,
  intersection_selected_sets,
  species
)
}
\arguments{
\item{significant_genes_by_comparison}{significant_genes_by_comparison}

\item{intersection_selected_sets}{intersection_selected_sets}

\item{species}{The latin binomial name of the species (used to make ensembl links)}
}
\value{
list of DT tables objects
}
\description{
gen_DT_lst
}
