% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{gen_selected_sets_xlsx}
\alias{gen_selected_sets_xlsx}
\title{gen_selected_sets_xlsx}
\usage{
gen_selected_sets_xlsx(lst, meta, file)
}
\arguments{
\item{lst}{list of data frames}

\item{meta}{metadata about the selected sets}

\item{file}{xlsx output file}
}
\value{
filename
}
\description{
gen_selected_sets_xlsx
}
