% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{gen_upset_plot}
\alias{gen_upset_plot}
\title{gen_upset_plot}
\usage{
gen_upset_plot(lst, combinations, set_2_highlight)
}
\arguments{
\item{lst}{named list of items in sets}

\item{combinations}{possible set combinations}

\item{set_2_highlight}{the sets to highlight}
}
\value{
upsetjs plot
}
\description{
gen_upset_plot
}
