% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{get_intersection_genes}
\alias{get_intersection_genes}
\title{get_intersection_genes}
\usage{
get_intersection_genes(mat, sets)
}
\arguments{
\item{mat}{binary set membership matrix (ComplexHeatmap)}

\item{sets}{vector of names of sets for which you want the intersection}
}
\value{
character vector of gene names
}
\description{
get the genes at the intersection of selected sets
}
