% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{get_significant_genes_by_comparison_lst}
\alias{get_significant_genes_by_comparison_lst}
\title{get_significant_genes_by_comparison_lst}
\usage{
get_significant_genes_by_comparison_lst(significant_genes_by_comparison)
}
\arguments{
\item{significant_genes_by_comparison}{significant_genes_by_comparison}
}
\value{
named list
}
\description{
Extracts gene set lists from filtered dge table
}
