% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{lfcpal}
\alias{lfcpal}
\title{lfcpal}
\usage{
lfcpal(x)
}
\arguments{
\item{x}{a numeric vector log2 Fold Changes}
}
\value{
DT::styleInterval
}
\description{
generates a colour scale for log2 fold change values for a DT table
}
