% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{named_group_split}
\alias{named_group_split}
\title{named_group_split}
\usage{
named_group_split(.tbl, ...)
}
\arguments{
\item{.tbl}{a tibble}

\item{...}{the groups to split}
}
\value{
a named list of tibbles
}
\description{
named_group_split
}
