% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{pvalpal}
\alias{pvalpal}
\title{pvalpal}
\usage{
pvalpal(x)
}
\arguments{
\item{x}{a numeric vector of p-values}
}
\value{
DT::styleInterval
}
\description{
generates a colour scale for p-values for a DT table
}
