% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{results_table_DT}
\alias{results_table_DT}
\title{results_table_DT}
\usage{
results_table_DT(
  data,
  log2fc_range = c(-5, 5),
  pvalue_range = c(0, 1),
  padj_range = c(0, 1)
)
}
\arguments{
\item{data}{a tibble with the results from a differential gene expression
analysis (specifics analysis ...)}

\item{log2fc_range}{set the range for the scales of the log2fc colour scale
numeric vector length 2 Default: c(-5,5)}

\item{pvalue_range}{set the range for the scales of the p-value colour scale
numeric vector length 2 Default: c(0,1)}

\item{padj_range}{set the range for the scales of the adjusted p-value colour scale
numeric vector length 2 Default: c(0,1)}
}
\value{
DT::datatable
}
\description{
results_table_DT
}
