% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{set_list_2_combinations}
\alias{set_list_2_combinations}
\title{set_list_2_combinations}
\usage{
set_list_2_combinations(set_list, add_empty = TRUE)
}
\arguments{
\item{set_list}{a named list where the names are the sets and the values
vectors of things in those sets}

\item{add_empty}{should an option to include an empty set be included?
if true (the default) adds an item called none with the value of an empty
string}
}
\value{
a list of combinations of sets
}
\description{
note that set order in names is derived from size of the sets, biggest first
also a none set with an empty string as a value is added by default
}
