% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sets_fct_sets_functions.R
\name{set_selected_set_2_inputs}
\alias{set_selected_set_2_inputs}
\title{set_selected_set_2_inputs}
\usage{
set_selected_set_2_inputs(x, upset, venn)
}
\arguments{
\item{x}{current value of comparisons}

\item{upset}{names from selector 1}

\item{venn}{names from selector 2}
}
\value{
character
}
\description{
deals with two upsets plots being able to provide inputs to the set
selection
}
